requirements/development.txt
-r base.txt

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code quality
black==23.12.1
flake8==6.1.0
isort==5.13.2

# Development server enhancements
django-livereload-server==0.4
